17-mcp-ai-ops/servers/postgres-mcp/requirements.txt
mcp>=0.9.0
psycopg2-binary>=2.9.9
